refactor: create shared BarLabelListTopLeft
